[CRT-382] Refactor custom task errors
